requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
bcftools=1.9=h47928c2_2
biopython=1.78=py37h7f8727e_0
blas=1.0=mkl
bottleneck=1.3.2=py37heb32a55_1
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py37h06a4308_2
curl=7.61.0=h84994c4_0
intel-openmp=2021.4.0=h06a4308_3561
ld_impl_linux-64=2.35.1=h7274673_9
libcurl=7.61.0=h1ad7b7a_0
libdeflate=1.0=h14c3975_1
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc=7.2.0=h69d50b8_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libssh2=1.8.0=h9cfc8f7_4
libstdcxx-ng=9.3.0=hd4cf53a_17
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.3=h7f8727e_2
numexpr=2.8.1=py37h6abb31d_0
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
openssl=1.0.2u=h7b6447c_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.4=py37h8c16a72_0
pip=21.2.2=py37h06a4308_0
primer3-py=0.6.0=py37h516909a_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysam=0.15.3=py37hda2845c_1
python=3.7.0=h6e4f718_3
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
readline=7.0=h7b6447c_5
samtools=1.7=1
seqtk=1.3=h5bf99c6_3
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
tk=8.6.11=h1ccaba5_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
